telegram_bot/requirements.txt
fastapi>=0.110.0
psycopg2>=2.9.9
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.30
python-dotenv>=1.0.0
aiogram>=2.17.0